+++ tests/integration/test_dummy_integration.py
@@ -6,6 +6,7 @@ from pathlib import Path
 import pytest
 from helpers import PROTEUS_ROOT
 from pandas.testing import assert_frame_equal
+from matplotlib.testing.compare import compare_images
 
 from proteus import Proteus
 from proteus.utils.coupler import ReadHelpfileFromCSV
@@ -13,7 +14,7 @@ from proteus.utils.coupler import ReadHelpfileFromCSV
 @pytest.fixture
 def test_dummy_run():
     config_path = PROTEUS_ROOT / 'input' / 'dummy.toml'
-    ref_output = PROTEUS_ROOT / 'tests' / 'data' / 'integration' / 'dummy' #/ 'runtime_helpfile.csv'
+    ref_output = PROTEUS_ROOT / 'tests' / 'data' / 'integration' / 'dummy'
 
     runner = Proteus(config_path=config_path)
 
@@ -25,4 +26,14 @@ def test_dummy_run():
     hf_all_ref = ReadHelpfileFromCSV(ref_output)
 
     assert_frame_equal(hf_all, hf_all_ref, rtol=1e-3)
-    #assert 'Completed' in (output / 'status').read_text()
+
+def test_plot_dummy_integration(test_dummy_run):
+
+    test_plot = PROTEUS_ROOT / 'output' / 'dummy' / 'plot_global_lin.png'
+    ref_plot = PROTEUS_ROOT / 'tests' / 'data' / 'integration' / 'dummy' / 'plot_global_lin.png'
+    tolerance = 5
+
+    result = compare_images(test_plot, ref_plot, tol=tolerance)
+
+    # compare_images returns None if the images are the same within the tolerance
+    assert result is None, f"The two PNG files differ more than the allowed tolerance: {result}"
